requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
pytest-emoji
pytest-sugar
selenium
page_objects
PyYAML